Update 1.8.Tapete.cpp
agora funciona no caso limite de l=1000000

+++ 1.8.Tapete.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 int main()
 {
-    int l, n;
+    unsigned long int l, n;
     cin >> l >> n;
     cout << n-1+(l-n+1)*(l-n+1);
     return 0;
